backend/requirements.txt
# --------------------------
# SmartAttendance Backend
# --------------------------

# Core Flask framework
Flask==3.0.3
Flask-Cors==4.0.1
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.6.0

# Database (MongoDB)
pymongo==4.10.1
Flask-PyMongo==2.3.0        # optional convenience wrapper
mongoengine==0.29.1         # optional ODM layer (use one or both as needed)
dnspython==2.6.1            # required for MongoDB Atlas connection URIs

# Environment and config
python-dotenv==1.0.1

# QR Code generation
qrcode==7.4.2
Pillow==11.0.0

# Geolocation / GeoJSON support
geopy==2.4.1
geojson==3.1.0

# Security and utilities
bcrypt==4.2.0
Werkzeug==3.0.4
requests==2.32.3             # useful for external APIs

# Date/time and data validation
pytz==2024.1
marshmallow==3.22.0

# Testing and debugging
pytest==8.3.3
flask-testing==0.8.1
black==24.4.2                # optional: code formatting

# Production server (optional)
gunicorn==23.0.0

